chore: move test file
