+++ service/follow_service/query_follower_list_test.go
@@ -1,15 +1,14 @@
-package test
+package follow_service
 
 import (
 	"douyin-lite/repository"
-	"douyin-lite/service/follow_service"
 	"fmt"
 	"testing"
 )
 
 func TestQueryFollowerListInfo(t *testing.T) {
 	repository.Init()
-	followerListInfo, err := follow_service.QueryFollowerListInfo(1)
+	followerListInfo, err := QueryFollowerListInfo(1)
 	if err != nil {
 		panic(err)
 	}
