+++ service/user_service/login_user_test.go
@@ -1,15 +1,14 @@
-package test
+package user_service
 
 import (
 	"douyin-lite/repository"
-	"douyin-lite/service/user_service"
 	"fmt"
 	"testing"
 )
 
 func TestLoginUser(t *testing.T) {
 	repository.Init()
-	loginInfo, err := user_service.LoginUser("feyman", "ywuiqjqiq")
+	loginInfo, err := LoginUser("feyman", "ywuiqjqiq")
 	if err != nil {
 		panic(err)
 	}
